requirements.txt
torch
torchvision
numpy
PyYAML
tqdm
matplotlib
scipy
